EXTEND-8 : Add a listener service to manage extendable resource removal. Add an implementation of listener to remove extenders of removed resources.
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/extends/module-extend-opengraph@48511 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/extend/modules/opengraph/service/extender/OpengraphResourceExtender.java
@@ -38,12 +38,11 @@ import fr.paris.lutece.plugins.extend.modules.opengraph.business.config.Opengrap
 import fr.paris.lutece.plugins.extend.service.extender.AbstractResourceExtender;
 import fr.paris.lutece.plugins.extend.service.extender.config.IResourceExtenderConfigService;
 
-import org.apache.commons.lang.StringUtils;
-
 import javax.inject.Inject;
 import javax.inject.Named;
-
 import javax.servlet.http.HttpServletRequest;
+
+import org.apache.commons.lang.StringUtils;
 
 
 /**
@@ -115,6 +114,9 @@ public class OpengraphResourceExtender extends AbstractResourceExtender
     @Override
     public void doDeleteResourceAddOn( ResourceExtenderDTO extender )
     {
-        _configService.remove( extender.getIdExtender(  ) );
+        if ( extender.getIdExtender( ) > 0 )
+        {
+            _configService.remove( extender.getIdExtender( ) );
+        }
     }
 }
